# expression: implement vectorized evaluation for `builtinQuarterSig`