<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement vectorized evaluation for builtinQuaterSig , for #12103

### What is changed and how it works?

	$go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args "builtinQuarterSig"
	goos: darwin
	goarch: amd64
	pkg: github.com/pingcap/tidb/expression
	BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.00756 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinQuarterSig-VecBuiltinFunc-4         	   57664	     20740 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinQuarterSig-NonVecBuiltinFunc-4      	   27924	     42456 ns/op	       0 B/op	       0 allocs/op
	PASS
	ok  	github.com/pingcap/tidb/expression	3.115s

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test